Update POSPageHeaderView to accept isSelected state from outside
