+++ WooCommerce/Classes/POS/Presentation/ItemListView.swift
@@ -182,6 +182,7 @@ private extension ItemListView {
         var items = [
             POSPageHeaderItem(
                 title: Localization.productsTitle,
+                isSelected: selectedItemListType.isProducts,
                 action: {
                     displayItemListType(.products(search: searchTerm.isNotEmpty))
                 }
@@ -192,6 +193,7 @@ private extension ItemListView {
             items.append(
                 POSPageHeaderItem(
                     title: Localization.couponsTitle,
+                    isSelected: selectedItemListType.isCoupons,
                     action: {
                         displayItemListType(.coupons)
                     }
