+++ WooCommerce/Classes/POS/Presentation/Reusable Views/POSPageHeaderView.swift
@@ -16,11 +16,13 @@ struct POSPageHeaderItem: Identifiable {
     let id = UUID()
     let title: String
     let subtitle: String?
+    let isSelected: Bool
     let action: (() -> Void)?
 
-    init(title: String, subtitle: String? = nil, action: (() -> Void)? = nil) {
+    init(title: String, subtitle: String? = nil, isSelected: Bool, action: (() -> Void)? = nil) {
         self.title = title
         self.subtitle = subtitle
+        self.isSelected = isSelected
         self.action = action
     }
 }
@@ -41,15 +43,13 @@ struct POSPageHeaderView<TrailingContent: View>: View {
         backButtonConfiguration != nil
     }
 
-    @State private var selectedItemIndex: Int = 0
-
     init(
         title: String,
         subtitle: String? = nil,
         backButtonConfiguration: POSPageHeaderBackButtonConfiguration? = nil,
         @ViewBuilder trailingContent: () -> TrailingContent = { EmptyView() }
     ) {
-        self.items = [.init(title: title, subtitle: subtitle)]
+        self.items = [.init(title: title, subtitle: subtitle, isSelected: true)]
         self.backButtonConfiguration = backButtonConfiguration
         self.trailingContent = trailingContent()
     }
@@ -74,7 +74,6 @@ struct POSPageHeaderView<TrailingContent: View>: View {
                 ForEach(0..<items.count, id: \.self) { index in
                     VStack(alignment: .leading, spacing: Constants.titleSubtitleSpacing) {
                         Button(action: {
-                            selectedItemIndex = index
                             items[index].action?()
                         }) {
                             Text(items[index].title)
@@ -82,9 +81,9 @@ struct POSPageHeaderView<TrailingContent: View>: View {
                                 .lineLimit(1)
                                 .minimumScaleFactor(0.5)
                                 .dynamicTypeSize(...DynamicTypeSize.accessibility2)
-                                .foregroundColor(selectedItemIndex == index ? .posOnSurface : .posOnSurfaceVariantLowest)
+                                .foregroundColor(items[index].isSelected ? .posOnSurface : .posOnSurfaceVariantLowest)
                         }
-                        .disabled(selectedItemIndex == index)
+                        .disabled(items[index].isSelected)
                         .accessibilityElement()
                         .accessibilityAddTraits(items.count == 1 ? .isHeader : [.isHeader, .isButton])
                         .accessibilityLabel(items[index].title)
@@ -211,11 +210,12 @@ private enum Constants {
             .foregroundColor(.posOnSurface)
         }
 
+        @State var isProductsSelected: Bool = true
         // Header with two items and trailing content.
         POSPageHeaderView(
             items: [
-                .init(title: "Products"),
-                .init(title: "Coupons")
+                .init(title: "Products", isSelected: isProductsSelected),
+                .init(title: "Coupons", isSelected: !isProductsSelected)
             ]
         ) {
             HStack(spacing: 16) {
